pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agenticai"
version = "0.1.0"
description = "Agentic AI - Twilio + Gemini + OpenClaw Gateway Integration"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AgenticAI Team"}
]
dependencies = [
    "google-genai>=1.0.0",
    "twilio>=9.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "numpy>=1.26.0",
    "soxr>=0.3.0",
    "audioop-lts>=0.2.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "APScheduler>=3.10.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
agenticai = "agenticai.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
